requirements.txt
Flask==3.0.0
pymongo==4.6.0
python-dotenv==1.0.0
google-generativeai==0.5.4
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
textblob==0.17.1
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
Pillow==10.0.0
